src/functions/import_adobe_analytics/requirements.txt
google-cloud-bigquery==3.5.0
google-cloud-secret-manager==2.15.1
pandas==1.5.3
google-auth==2.16.0